+++ src/main/java/gov/usgs/wma/waterdata/collections/ObservationsController.java
@@ -59,9 +59,8 @@ public class ObservationsController extends BaseController {
 			@PathVariable(value = PARAM_COLLECTION_ID) String collectionId,
 			@PathVariable(value = PARAM_FEATURE_ID) String featureId,
 			HttpServletResponse response) {
-		collectionsParams.builder.clear();
 
-		String 	rtn = collectionsDao.getObsverationsJson(collectionsParams.builder.collectionId(collectionId)
+		String 	rtn = collectionsDao.getObsverationsJson(collectionsParams.builder().collectionId(collectionId)
 			.featureId(featureId).build());
 		if (rtn == null) {
 			response.setStatus(HttpStatus.NOT_FOUND.value());
@@ -92,9 +91,8 @@ public class ObservationsController extends BaseController {
 			@PathVariable(value = PARAM_COLLECTION_ID) String collectionId,
 			@PathVariable(value = PARAM_FEATURE_ID) String featureId,
 			HttpServletResponse response) {
-		collectionsParams.builder.clear();
 
-		String 	rtn = collectionsDao.getDiscreteDataObsverationsJson(collectionsParams.builder.collectionId(collectionId)
+		String 	rtn = collectionsDao.getDiscreteDataObsverationsJson(collectionsParams.builder().collectionId(collectionId)
 			.featureId(featureId).build());
 		if (rtn == null) {
 			response.setStatus(HttpStatus.NOT_FOUND.value());