+++ src/test/java/gov/usgs/wma/waterdata/collections/FeaturesFilterIT.java
@@ -8,6 +8,8 @@ import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
+import com.github.springtestdbunit.annotation.DatabaseSetup;
+
 import java.io.IOException;
 
 import static org.hamcrest.CoreMatchers.equalTo;
@@ -15,6 +17,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
+@DatabaseSetup("classpath:/testData/monitoringLocation/")
 public class FeaturesFilterIT extends BaseCollectionsIT {
     @Autowired
     private TestRestTemplate restTemplate;
@@ -30,7 +33,7 @@ public class FeaturesFilterIT extends BaseCollectionsIT {
 
     @Test
     public void countryTestNotFound() throws IOException {
-        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?country=MX",
+        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?countryFIPS=MX",
             String.class);
         assertThat(rtn.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
         assertEquals(ogc404Payload, rtn.getBody());
@@ -38,14 +41,14 @@ public class FeaturesFilterIT extends BaseCollectionsIT {
 
     @Test
     public void countryTestInvalid() throws IOException {
-        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?country=MXUS",
+        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?countryFIPS=MXUS",
             String.class);
         assertThat(rtn.getStatusCode(), equalTo(HttpStatus.BAD_REQUEST));
     }
 
     @Test
     public void stateTest() throws IOException {
-        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?state=26",
+        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?stateFIPS=US:26",
             String.class);
 
         assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
@@ -55,7 +58,7 @@ public class FeaturesFilterIT extends BaseCollectionsIT {
 
     @Test
     public void stateTestNotFound() throws IOException {
-        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?state=57",
+        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?stateFIPS=US:99",
             String.class);
         assertThat(rtn.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
         assertEquals(ogc404Payload, rtn.getBody());
@@ -63,14 +66,14 @@ public class FeaturesFilterIT extends BaseCollectionsIT {
 
     @Test
     public void stateTestInvalid() throws IOException {
-        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?state=57000",
+        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?stateFIPS=57000",
             String.class);
         assertThat(rtn.getStatusCode(), equalTo(HttpStatus.BAD_REQUEST));
     }
 
     @Test
     public void countyTest() throws IOException {
-        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?county=05051",
+        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?countyFIPS=US:05:051",
             String.class);
 
         assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
@@ -80,7 +83,7 @@ public class FeaturesFilterIT extends BaseCollectionsIT {
 
     @Test
     public void countyTestNotFound() throws IOException {
-        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?county=06071",
+        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?countyFIPS=US:00:051",
             String.class);
         assertThat(rtn.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
         assertEquals(ogc404Payload, rtn.getBody());
@@ -88,14 +91,14 @@ public class FeaturesFilterIT extends BaseCollectionsIT {
 
     @Test
     public void countyTestInvalid() throws IOException {
-        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?county=Hartley",
+        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?countyFIPS=Hartley",
             String.class);
         assertThat(rtn.getStatusCode(), equalTo(HttpStatus.BAD_REQUEST));
     }
 
     @Test
-    public void hydrologicalUnitsTest() throws IOException {
-        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?hydrologicalUnit=040103020107",
+    public void hydrologicUnitsTest() throws IOException {
+        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?hydrologicUnit=040103020107",
             String.class);
 
         assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
@@ -103,17 +106,28 @@ public class FeaturesFilterIT extends BaseCollectionsIT {
         doJsonCompare(rtn, compareFile);
     }
 
+	@Test
+	public void hydrologicUnitsMultiValuesTest() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity(
+				"/collections/monitoring-locations/items?hydrologicUnit=040103020107&hydrologicUnit=080401010805",
+				String.class);
+
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
+		String compareFile = "featureCollection/monitoring-locations/hydrological_unit_multi_values.json";
+		doJsonCompare(rtn, compareFile);
+	}
+
     @Test
-    public void hydrologicalUnitsTestNotFound() throws IOException {
-        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?hydrologicalUnit=000000000000",
+    public void hydrologicUnitsTestNotFound() throws IOException {
+        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?hydrologicUnit=000000000000",
             String.class);
         assertThat(rtn.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
         assertEquals(ogc404Payload, rtn.getBody());
     }
 
     @Test
-    public void hydrologicalUnitsTestInvalid() throws IOException {
-        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?hydrologicalUnit=INVALID",
+    public void hydrologicUnitsTestInvalid() throws IOException {
+        ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?hydrologicUnit=INVALID",
             String.class);
         assertThat(rtn.getStatusCode(), equalTo(HttpStatus.BAD_REQUEST));
     }
@@ -126,6 +140,16 @@ public class FeaturesFilterIT extends BaseCollectionsIT {
         String compareFile = "featureCollection/monitoring-locations/monitoring_location_type_well.json";
         doJsonCompare(rtn, compareFile);
     }
+
+	@Test
+	public void monitoringLocationTypeMultiValuesTest() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity(
+				"/collections/monitoring-locations/items?monitoringLocationType=Well&monitoringLocationType=Stream",
+				String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
+		String compareFile = "featureCollection/monitoring-locations/monitoring_location_type_multi_values.json";
+		doJsonCompare(rtn, compareFile);
+	}
 
     @Test
     public void monitoringLocationTypeTestNotFound() throws IOException {
@@ -172,9 +196,8 @@ public class FeaturesFilterIT extends BaseCollectionsIT {
     @Test
     public void monitoringLocationNumberTest() throws IOException {
         ResponseEntity<String> rtn = restTemplate.getForEntity(
-            "/collections/monitoring-locations/items?monitoringLocationNumber=USGS-343204093005501",
+            "/collections/monitoring-locations/items?monitoringLocationNumber=343204093005501",
             String.class);
-
         assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
         String compareFile = "featureCollection/monitoring-locations/monitoring_location_number_USGS-343204093005501.json";
         doJsonCompare(rtn, compareFile);