+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -   Performance improvement on end point: /collections/{collectionId}/items/{featureId}/observations/statistical-time-series
 -   Unapproved time series data older than 1095 days is not returned from the services (USGS rule)
 -   Added filtering to /collections/{collectionId}/items
+-   Added nationalAquiferCode to feature properties
+-   Deprecated end point: /monitoring-location/{monitoringLocationId}
 
 ## [0.6.0](https://github.com/USGS/time-series-services/compare/time-series-services-0.5.0...time-series-services-0.6.0) - 2020-04-29
 ### Changed